+++ src/components/dashboard/currencyGraph.test.js
@@ -3,8 +3,8 @@ import { expect } from 'chai';
 import { Line as LineChart } from 'react-chartjs-2';
 import sinon from 'sinon';
 import React from 'react';
-import { mountWithContext } from './../../../test/utils/mountHelpers';
-import { prepareStore } from '../../../test/utils/applicationInit';
+import { mountWithContext } from '../../../test/unit-test-utils/mountHelpers';
+import { prepareStore } from '../../../test/unit-test-utils/applicationInit';
 import liskServiceApi from '../../utils/api/liskService';
 
 import liskServiceReducer from '../../store/reducers/liskService';
