+++ src/components/voting/voting.test.js
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { mount } from 'enzyme';
 import PropTypes from 'prop-types';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { prepareStore } from '../../../test/utils/applicationInit';
+import { prepareStore } from '../../../test/unit-test-utils/applicationInit';
 import peersReducer from '../../store/reducers/peers';
 import accountReducer from '../../store/reducers/account';
 import votingReducer from '../../store/reducers/voting';
