+++ src/components/votingListView/votingListView.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import VotingListView from './votingListView';
-import { mountWithContext } from './../../../test/utils/mountHelpers';
+import { mountWithContext } from '../../../test/unit-test-utils/mountHelpers';
 import store from '../../store';
 import voteFilters from './../../constants/voteFilters';
